+++ oshi-core/src/main/java/oshi/hardware/platform/mac/MacSensors.java
@@ -24,7 +24,7 @@
 package oshi.hardware.platform.mac;
 
 import oshi.hardware.common.AbstractSensors;
-import oshi.jna.platform.mac.IOKit;
+import oshi.jna.platform.mac.IOKit.IOConnect;
 import oshi.util.platform.mac.SmcUtil;
 
 /**
@@ -39,9 +39,9 @@ public class MacSensors extends AbstractSensors {
 
     @Override
     public double queryCpuTemperature() {
-        SmcUtil.smcOpen();
-        double temp = SmcUtil.smcGetFloat(IOKit.SMC_KEY_CPU_TEMP, 50);
-        SmcUtil.smcClose();
+        IOConnect conn = SmcUtil.smcOpen();
+        double temp = SmcUtil.smcGetFloat(conn, SmcUtil.SMC_KEY_CPU_TEMP);
+        SmcUtil.smcClose(conn);
         if (temp > 0d) {
             return temp;
         }
@@ -51,23 +51,23 @@ public class MacSensors extends AbstractSensors {
     @Override
     public int[] queryFanSpeeds() {
         // If we don't have fan # try to get it
-        SmcUtil.smcOpen();
+        IOConnect conn = SmcUtil.smcOpen();
         if (this.numFans == 0) {
-            this.numFans = (int) SmcUtil.smcGetLong(IOKit.SMC_KEY_FAN_NUM, 50);
+            this.numFans = (int) SmcUtil.smcGetLong(conn, SmcUtil.SMC_KEY_FAN_NUM);
         }
         int[] fanSpeeds = new int[this.numFans];
         for (int i = 0; i < this.numFans; i++) {
-            fanSpeeds[i] = (int) SmcUtil.smcGetFloat(String.format(IOKit.SMC_KEY_FAN_SPEED, i), 50);
+            fanSpeeds[i] = (int) SmcUtil.smcGetFloat(conn, String.format(SmcUtil.SMC_KEY_FAN_SPEED, i));
         }
-        SmcUtil.smcClose();
+        SmcUtil.smcClose(conn);
         return fanSpeeds;
     }
 
     @Override
     public double queryCpuVoltage() {
-        SmcUtil.smcOpen();
-        double volts = SmcUtil.smcGetFloat(IOKit.SMC_KEY_CPU_VOLTAGE, 50) / 1000d;
-        SmcUtil.smcClose();
+        IOConnect conn = SmcUtil.smcOpen();
+        double volts = SmcUtil.smcGetFloat(conn, SmcUtil.SMC_KEY_CPU_VOLTAGE) / 1000d;
+        SmcUtil.smcClose(conn);
         return volts;
     }
 }