+++ oshi-core/src/main/java/oshi/util/platform/mac/SmcUtil.java
@@ -32,24 +32,88 @@ import java.util.concurrent.ConcurrentHashMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.sun.jna.platform.mac.SystemB; // NOSONAR
-import com.sun.jna.ptr.IntByReference;
+import com.sun.jna.NativeLong;
+import com.sun.jna.Structure;
+import com.sun.jna.Structure.FieldOrder;
+import com.sun.jna.ptr.NativeLongByReference;
+import com.sun.jna.ptr.PointerByReference;
 
 import oshi.jna.platform.mac.IOKit;
+import oshi.jna.platform.mac.IOKitUtil;
+import oshi.jna.platform.mac.SystemB;
 import oshi.jna.platform.mac.IOKit.IOConnect;
-import oshi.jna.platform.mac.IOKit.SMCKeyData;
-import oshi.jna.platform.mac.IOKit.SMCKeyDataKeyInfo;
-import oshi.jna.platform.mac.IOKit.SMCVal;
+import oshi.jna.platform.mac.IOKit.IOService;
 import oshi.util.ParseUtil;
-import oshi.util.Util;
 
 /**
  * Provides access to SMC calls on OS X
  */
 public class SmcUtil {
+    /**
+     * Holds the return value of SMC version query.
+     */
+    @FieldOrder({ "major", "minor", "build", "reserved", "release" })
+    public static class SMCKeyDataVers extends Structure {
+        public byte major;
+        public byte minor;
+        public byte build;
+        public byte reserved;
+        public short release;
+
+    }
+
+    /**
+     * Holds the return value of SMC pLimit query.
+     */
+    @FieldOrder({ "version", "length", "cpuPLimit", "gpuPLimit", "memPLimit" })
+    public static class SMCKeyDataPLimitData extends Structure {
+        public short version;
+        public short length;
+        public int cpuPLimit;
+        public int gpuPLimit;
+        public int memPLimit;
+    }
+
+    /**
+     * Holds the return value of SMC KeyInfo query.
+     */
+    @FieldOrder({ "dataSize", "dataType", "dataAttributes" })
+    public static class SMCKeyDataKeyInfo extends Structure {
+        public int dataSize;
+        public int dataType;
+        public byte dataAttributes;
+    }
+
+    /**
+     * Holds the return value of SMC query.
+     */
+    @FieldOrder({ "key", "vers", "pLimitData", "keyInfo", "result", "status", "data8", "data32", "bytes" })
+    public static class SMCKeyData extends Structure {
+        public int key;
+        public SMCKeyDataVers vers;
+        public SMCKeyDataPLimitData pLimitData;
+        public SMCKeyDataKeyInfo keyInfo;
+        public byte result;
+        public byte status;
+        public byte data8;
+        public int data32;
+        public byte[] bytes = new byte[32];
+    }
+
+    /**
+     * Holds an SMC value
+     */
+    @FieldOrder({ "key", "dataSize", "dataType", "bytes" })
+    public static class SMCVal extends Structure {
+        public byte[] key = new byte[5];
+        public int dataSize;
+        public byte[] dataType = new byte[5];
+        public byte[] bytes = new byte[32];
+    }
+
     private static final Logger LOG = LoggerFactory.getLogger(SmcUtil.class);
 
-    private static IOConnect conn = new IOConnect();
+    private static IOKit IO = IOKit.INSTANCE;
 
     /**
      * Map for caching info retrieved by a key necessary for subsequent calls.
@@ -63,55 +127,64 @@ public class SmcUtil {
     private static final byte[] DATATYPE_FPE2 = ParseUtil.stringToByteArray("fpe2", 5);
     private static final byte[] DATATYPE_FLT = ParseUtil.stringToByteArray("flt ", 5);
 
+    public static final String SMC_KEY_FAN_NUM = "FNum";
+    public static final String SMC_KEY_FAN_SPEED = "F%dAc";
+    public static final String SMC_KEY_CPU_TEMP = "TC0P";
+    public static final String SMC_KEY_CPU_VOLTAGE = "VC0C";
+
+    public static final byte SMC_CMD_READ_BYTES = 5;
+    public static final byte SMC_CMD_READ_KEYINFO = 9;
+    public static final int KERNEL_INDEX_SMC = 2;
+
     private SmcUtil() {
     }
 
     /**
-     * Open a connection to SMC
+     * Open a connection to SMC.
      *
-     * @return 0 if successful, nonzero if failure
+     * @return The connection if successful, null if failure
      */
-    public static int smcOpen() {
-        int service = IOKitUtil.getMatchingService("AppleSMC");
-        if (service == 0) {
-            LOG.error("Error: no SMC found");
-            return 1;
-        }
-
-        int result = IOKit.INSTANCE.IOServiceOpen(service, SystemB.INSTANCE.mach_task_self(), 0, conn);
-        IOKit.INSTANCE.IOObjectRelease(service);
-        if (result != 0) {
-            if (LOG.isErrorEnabled()) {
-                LOG.error(String.format("Error: IOServiceOpen() = 0x%08x", result));
+    public static IOConnect smcOpen() {
+        IOService smcService = IOKitUtil.getMatchingService("AppleSMC");
+        if (smcService != null) {
+            PointerByReference connPtr = new PointerByReference();
+            int result = IO.IOServiceOpen(smcService, SystemB.INSTANCE.mach_task_self(), 0, connPtr);
+            smcService.release();
+            if (result == 0) {
+                return new IOConnect(connPtr.getValue());
+            } else if (LOG.isErrorEnabled()) {
+                LOG.error(String.format("Unable to open connection to AppleSMC service. Error: 0x%08x", result));
             }
-            return result;
+        } else {
+            LOG.error("Unable to locate AppleSMC service");
         }
-        // Delay to improve success of next query
-        Util.sleep(5);
-        return 0;
+        return null;
     }
 
     /**
-     * Close connection to SMC
+     * Close connection to SMC.
+     *
+     * @param conn
+     *            The connection
      *
      * @return 0 if successful, nonzero if failure
      */
-    public static int smcClose() {
-        return IOKit.INSTANCE.IOServiceClose(conn.getValue());
+    public static int smcClose(IOConnect conn) {
+        return IO.IOServiceClose(conn);
     }
 
     /**
      * Get a value from SMC which is in a floating point datatype (SP78, FPE2, FLT)
      *
+     * @param conn
+     *            The connection
      * @param key
      *            The key to retrieve
-     * @param retries
-     *            Number of times to retry the key
      * @return Double representing the value
      */
-    public static double smcGetFloat(String key, int retries) {
+    public static double smcGetFloat(IOConnect conn, String key) {
         SMCVal val = new SMCVal();
-        int result = smcReadKey(key, val, retries);
+        int result = smcReadKey(conn, key, val);
         if (result == 0 && val.dataSize > 0) {
             if (Arrays.equals(val.dataType, DATATYPE_SP78) && val.dataSize == 2) {
                 // First bit is sign, next 7 bits are integer portion, last 8 bits are
@@ -132,15 +205,15 @@ public class SmcUtil {
     /**
      * Get a 64-bit integer value from SMC
      *
+     * @param conn
+     *            The connection
      * @param key
      *            The key to retrieve
-     * @param retries
-     *            Number of times to retry the key
      * @return Long representing the value
      */
-    public static long smcGetLong(String key, int retries) {
+    public static long smcGetLong(IOConnect conn, String key) {
         SMCVal val = new SMCVal();
-        int result = smcReadKey(key, val, retries);
+        int result = smcReadKey(conn, key, val);
         if (result == 0) {
             return ParseUtil.byteArrayToLong(val.bytes, val.dataSize);
         }
@@ -151,22 +224,23 @@ public class SmcUtil {
     /**
      * Get cached keyInfo if it exists, or generate new keyInfo
      *
+     * @param conn
+     *            The connection
      * @param inputStructure
      *            Key data input
      * @param outputStructure
      *            Key data output
      * @return 0 if successful, nonzero if failure
      */
-    public static int smcGetKeyInfo(SMCKeyData inputStructure, SMCKeyData outputStructure) {
+    public static int smcGetKeyInfo(IOConnect conn, SMCKeyData inputStructure, SMCKeyData outputStructure) {
         if (keyInfoCache.containsKey(inputStructure.key)) {
             SMCKeyDataKeyInfo keyInfo = keyInfoCache.get(inputStructure.key);
             outputStructure.keyInfo.dataSize = keyInfo.dataSize;
             outputStructure.keyInfo.dataType = keyInfo.dataType;
             outputStructure.keyInfo.dataAttributes = keyInfo.dataAttributes;
         } else {
-            inputStructure.data8 = IOKit.SMC_CMD_READ_KEYINFO;
-            Util.sleep(4);
-            int result = smcCall(IOKit.KERNEL_INDEX_SMC, inputStructure, outputStructure);
+            inputStructure.data8 = SMC_CMD_READ_KEYINFO;
+            int result = smcCall(conn, KERNEL_INDEX_SMC, inputStructure, outputStructure);
             if (result != 0) {
                 return result;
             }
@@ -182,55 +256,41 @@ public class SmcUtil {
     /**
      * Read a key from SMC
      *
+     * @param conn
+     *            The connection
      * @param key
      *            Key to read
      * @param val
      *            Structure to receive the result
-     * @param retries
-     *            Number of attempts to try
      * @return 0 if successful, nonzero if failure
      */
-    public static int smcReadKey(String key, SMCVal val, int retries) {
+    public static int smcReadKey(IOConnect conn, String key, SMCVal val) {
         SMCKeyData inputStructure = new SMCKeyData();
         SMCKeyData outputStructure = new SMCKeyData();
 
         inputStructure.key = (int) ParseUtil.strToLong(key, 4);
-        int result;
-        // These calls frequently result in kIOReturnIPCError so retry multiple
-        // times with small delay
-        int retry = 0;
-        do {
-            result = smcGetKeyInfo(inputStructure, outputStructure);
-            if (result != 0) {
-                continue;
-            }
-
+        int result = smcGetKeyInfo(conn, inputStructure, outputStructure);
+        if (result == 0) {
             val.dataSize = outputStructure.keyInfo.dataSize;
             val.dataType = ParseUtil.longToByteArray(outputStructure.keyInfo.dataType, 4, 5);
 
             inputStructure.keyInfo.dataSize = val.dataSize;
-            inputStructure.data8 = IOKit.SMC_CMD_READ_BYTES;
+            inputStructure.data8 = SMC_CMD_READ_BYTES;
 
-            Util.sleep(4);
-            result = smcCall(IOKit.KERNEL_INDEX_SMC, inputStructure, outputStructure);
-            // If we got success, exit!
+            result = smcCall(conn, KERNEL_INDEX_SMC, inputStructure, outputStructure);
             if (result == 0) {
-                break;
+                System.arraycopy(outputStructure.bytes, 0, val.bytes, 0, val.bytes.length);
+                return 0;
             }
-        } while (++retry < retries);
-        // If we errored out return code
-        if (result != 0) {
-            return result;
         }
-
-        System.arraycopy(outputStructure.bytes, 0, val.bytes, 0, val.bytes.length);
-        // Success
-        return 0;
+        return result;
     }
 
     /**
      * Call SMC
      *
+     * @param conn
+     *            The connection
      * @param index
      *            Kernel index
      * @param inputStructure
@@ -239,8 +299,8 @@ public class SmcUtil {
      *            Key data output
      * @return 0 if successful, nonzero if failure
      */
-    public static int smcCall(int index, SMCKeyData inputStructure, SMCKeyData outputStructure) {
-        return IOKit.INSTANCE.IOConnectCallStructMethod(conn.getValue(), index, inputStructure, inputStructure.size(),
-                outputStructure, new IntByReference(outputStructure.size()));
+    public static int smcCall(IOConnect conn, int index, SMCKeyData inputStructure, SMCKeyData outputStructure) {
+        return IO.IOConnectCallStructMethod(conn, index, inputStructure, new NativeLong(inputStructure.size()),
+                outputStructure, new NativeLongByReference(new NativeLong(outputStructure.size())));
     }
 }